requirements.txt
absl-py==2.1.0
addict==2.4.0
aioftp==0.25.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asdf==4.1.0
asdf-astropy==0.6.1
asdf_coordinates_schemas==0.3.0
asdf_standard==1.1.1
asdf_transform_schemas==0.5.0
astropy==6.0.1
astropy-iers-data==0.2025.3.10.0.29.26
astropy_healpix==1.0.3
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
Cartopy==0.23.0
cdflib==1.3.4
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
configmypy==0.2.0
contourpy==1.3.0
coverage==7.6.12
cycler==0.12.1
Cython==3.0.12
dash==3.0.2
dask==2024.8.0
debugpy==1.8.11
decorator==5.1.1
-e git+https://github.com/deepsphere/deepsphere-cosmo-tf2.git@810190a020e0ba8e6568ba5175c57dd3ed5b8a20#egg=deepsphere
DeepXDE==1.13.2
defusedxml==0.7.1
distlib==0.3.9
docker-pycreds==0.4.0
drms==0.6.4
einshape==1.0
etils==1.5.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
flatbuffers==25.2.10
flax==0.8.5
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
Glymur==0.14.0.post1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5netcdf==1.6.1
h5py==3.12.1
healpy==1.17.3
HelioPy==0.15.3
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
identify==2.6.9
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.18.1
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
libigl==2.5.1
lightning-utilities==0.14.3
llvmlite==0.43.0
locket==1.0.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpl-animators==1.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
neuraloperator==1.0.2
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.60.0
numcodecs==0.12.1
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.6.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parfive==2.1.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pickle5==0.0.11
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
pooch==1.8.2
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==5.29.3
psipy==0.4.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyerfa==2.0.1.5
pyflakes==3.2.0
pyflann3==1.8.4.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyGSP @ git+https://github.com/epfl-lts2/pygsp@a8019a9db1f3813a2f37915596aafbc261c83dcf
pyhdf==0.11.6
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
PyOpenGL==3.1.9
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyqtgraph==0.10.0
pyquaternion==0.9.9
PyRect==0.2.0
PyScreeze==1.0.1
pyshp==2.3.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytorch-msssim==1.0.0
pytweening==1.2.0
pytz==2025.1
pyvista==0.44.2
PyYAML==6.0.2
pyzmq==26.2.0
quaternionic==1.0.15
referencing==0.36.2
reproject==0.13.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
rubicon-objc==0.5.0
s3pipe==1.3.0
scikit-image==0.24.0
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.13.1
scooby==0.10.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.25.0
setproctitle==1.3.5
sewar==0.4.6
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sparselinear==0.0.5
spherical==1.0.15
spiceypy==6.0.0
spinsfast==2022.4.10
SQLAlchemy==2.0.40
stack-data==0.6.3
sunpy==5.1.5
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorly==0.9.0
tensorly-torch==0.5.0
tensorstore==0.1.69
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torch-geometric==2.6.1
torch_harmonics==0.7.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchmetrics==1.7.3
torchsummary==1.5.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.4
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.29.3
vtk==9.3.1
wandb==0.19.9
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2024.7.0
yarl==1.18.3
zarr==2.18.2
zeep==4.3.1
zipp==3.21.0